requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
h11==0.12.0
idna==3.2
mailchimp-marketing==3.0.66
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
